Many cleanups
HeapSizeEstimator renamed to ClassSizeEstimator
MeasurableUtils renamed to InstanceSizeEstimator

HeapSizeEstimatorTest is now abstract, with two subclasses: ClassSizeEstimatorTest and InstanceSizeEstimatorTest.
