+++ gradle/spotbugs/exclude.xml
@@ -70,19 +70,12 @@
     </Or>
   </Match>
   <Match>
-    <!--Explicit GC is intentional in some rare testing contexts-->
-    <Bug pattern="DM_GC"/>
-    <Or>
-      <Class name="com.linkedin.venice.memory.HeapSizeEstimatorTest"/>
-    </Or>
-  </Match>
-  <Match>
     <!--These classes pass null intentionally for testing purposes-->
     <Bug pattern="NP_NONNULL_PARAM_VIOLATION"/>
     <Or>
       <Class name="com.linkedin.venice.router.api.TestHostFinder"/>
       <Class name="com.linkedin.davinci.StoreBackendTest"/>
-      <Class name="com.linkedin.venice.memory.HeapSizeEstimatorTest"/>
+      <Class name="com.linkedin.venice.memory.ClassSizeEstimatorTest"/>
     </Or>
   </Match>
   <Match>
