+++ internal/venice-common/src/main/java/com/linkedin/venice/writer/CompletableFutureCallback.java
@@ -1,6 +1,6 @@
 package com.linkedin.venice.writer;
 
-import static com.linkedin.venice.memory.HeapSizeEstimator.getClassOverhead;
+import static com.linkedin.venice.memory.ClassSizeEstimator.getClassOverhead;
 
 import com.linkedin.venice.memory.Measurable;
 import com.linkedin.venice.pubsub.api.PubSubProduceResult;
